pyproject.toml
[tool.poetry]
name = "matching"
version = "0.1.0"
description = "Web UI to annotate db data pairs similarity"
authors = ["Kevin Samuel <conctact@kevinsamuel.email>"]
readme = "README.md"
homepage = "https://github.com/ksamuel/matching/"
repository = "https://github.com/ksamuel/matching.git"

[tool.poetry.dependencies]
python = "^3.6"
django = "==3.1.7"
django-cache-url = "==0.8.0"
djangorestframework = "==3.12.4"
gunicorn = "==20.1.0"
lxml = "==4.6.3"
psycopg2 = "==2.8.6"
python-dotenv = "==0.17.1"
redis = "==3.5.3"
sqlalchemy = "==1.4.12"
django_extensions = "^3.1"
django-cors-headers = "^3.7"
dj-database-url = "^0.5.0"
environs = {version = "^9.3.2", extras = ["django"]}
django-redis = "^5.0.0"
Werkzeug = "^2.0.1"

[tool.poetry.dev-dependencies]
django-stubs = "^1.8"
jupyter = "^1.0"
mypy = "^0.812.0"
pylint = "^2.8"
pylint-django = "^2.4"
pylint-plugin-utils = "^0.6.0"
pytest = "^6.2"
pytest-django = "^4.2"
poetry = "^1.1"
pytest-clarity = {version = "^0.1.0-alpha.1", allow-prereleases = true}
black = {version = "^18.3-alpha.0", allow-prereleases = true}
rich = "^10.2"
bettererrormessages = "^0.4.0"
friendly = "^0.3.65"
doit = "^0.33.1"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
